fix: update seeding script with donors, volunteers, checkin, content (#174)
* chore: update seeding script

fix: emails

* chore: update readme with new seeding credentials

* fix: cascade drop tables

* fix seeding
